[DUDI:] changed all date format and added delete-post authorizartion

+++ controllers/posts-controller.js
@@ -13,10 +13,24 @@ const getAllPosts = (req, res, next) => {
 
 const deletePostById = (req, res, next) => {
     console.log('POST request in user.<id>.post');
-    const postId = req.body.postId;
-    dataManager.deleteData("posts", postId);
-
-    res.json({Message: "deleted postId " + postId});
+    // const postId = req.body.postId;
+    const { userId, postId } = req.body;
+    const userPost = dataManager.getData("posts").find((post) => post.id === postId);
+    if (post && post.uId === userId) {
+        dataManager.deleteData("posts", postId);
+        res.json({Message: "deleted postId " + postId});
+    }
+    else {
+        const message;
+        if (!post) {
+            message = `post by id ${postId} not found`
+            return next(new HttpError(message, 400));
+        }
+        else {
+            message = `Not authorize: user is not post creator`
+            return next(new HttpError(message, 403));
+        }
+    }
 }
 
 exports.getAllPosts = getAllPosts;
